// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-preflight-source-body.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingOperationUnitBody } from '../models/linking-operation-unit-body';
import { LinkingSourceDescriptorBody } from '../models/linking-source-descriptor-body';
export interface LinkingPreflightSourceBody {
  automaticWordMatchesEnabled: boolean | null;
  contributionMode: string | null;
  descriptor: LinkingSourceDescriptorBody | null;
  orderValue: number | null;
  units: Array<LinkingOperationUnitBody> | null;
}